backend/requirements.txt
fastapi
uvicorn[standard]
spotipy
google-cloud-language
python-dotenv
pydantic
vaderSentiment
